person-api/requirements.txt
python-jose
python-dotenv
six
everett
credstash
